+++ js/conf/Debug.js
@@ -1,16 +1,18 @@
 // Set prod to true when releasing
-// _prod = true; 
-_prod = false; 
+_prod = true; 
+// _prod = false; 
 
 Debug = {
-  debug: true,
+  init: true,
+  debug: false,
   keyboard: true,
   debugResizer: true,
   debugArDetect: true,
-  debugStorage: true,
-  comms: true,
+  debugStorage: false,
+  comms: false,
   // showArDetectCanvas: true,
-  flushStoredSettings: false,
+  flushStoredSettings: true,
+  // flushStoredSettings: false,
   playerDetectDebug: true,
   periodic: true,
   videoRescan: true,