+++ js/lib/VideoData.js
@@ -1,10 +1,10 @@
 class VideoData {
   
-  constructor(video){
+  constructor(video, settings, pageInfo){
     this.arSetupComplete = false;
     this.video = video;
     this.destroyed = false;
-
+    this.settings = settings;
     // POZOR: VRSTNI RED JE POMEMBEN (arDetect mora bit zadnji)
     // NOTE: ORDERING OF OBJ INITIALIZATIONS IS IMPORTANT (arDetect needs to go last)    
     this.player = new PlayerData(this);
@@ -13,15 +13,26 @@ class VideoData {
     this.arDetector = new ArDetector(this);  // this starts Ar detection. needs optional parameter that prevets ardetdctor from starting
     // player dimensions need to be in:
     // this.player.dimensions
+    this.pageInfo = pageInfo;
+    this.vdid = (Math.random()*100).toFixed();
+    if (Debug.init) {
+      console.log("[VideoData::ctor] Created videoData with vdid", this.vdid);
+    }
   }
 
   firstTimeArdInit(){
+    if(this.destroyed) {
+      throw {error: 'VIDEO_DATA_DESTROYED', data: {videoData: this}};
+    }
     if(! this.arSetupComplete){
       this.arDetector = new ArDetector(this);
     }
   }
 
   initArDetection() {
+    if(this.destroyed) {
+      throw {error: 'VIDEO_DATA_DESTROYED', data: {videoData: this}};
+    }
     if(this.arDetector){
       this.arDetector.init();
     }
@@ -32,14 +43,27 @@ class VideoData {
   }
   
   startArDetection() {
+    if(this.destroyed) {
+      throw {error: 'VIDEO_DATA_DESTROYED', data: {videoData: this}};
+    }
+    if(!this.arDetector) {
+      this.arDetector.init();
+    }
     this.arDetector.start();
   }
 
   stopArDetection() {
-    this.arDetector.stop();
+    if (this.arDetector) {
+      this.arDetector.stop();
+    }
   }
 
   destroy() {
+    if(Debug.debug || Debug.init){ 
+      console.log(`[VideoData::destroy] <vdid:${this.vdid}> received destroy command`);
+    }
+
+    this.pause();
     this.destroyed = true;
     if(this.arDetector){
       this.arDetector.stop();
@@ -49,9 +73,11 @@ class VideoData {
     if(this.resizer){
       this.resizer.destroy();
     }
+    this.resizer = null;
     if(this.player){
-      player.destroy();
+      this.player.destroy();
     }
+    this.player = null;
     this.video = null;
   }
 
@@ -69,6 +95,9 @@ class VideoData {
   }
 
   resume(){
+    if(this.destroyed) {
+      throw {error: 'VIDEO_DATA_DESTROYED', data: {videoData: this}};
+    }
     this.paused = false;
     try {
       this.resizer.start();
@@ -95,6 +124,29 @@ class VideoData {
     this.resizer.setAr(ar, lastAr);
   }
 
+  resetAr() {
+    this.resizer.reset();
+  }
+
+  panHandler(event) {
+    if(this.destroyed) {
+      throw {error: 'VIDEO_DATA_DESTROYED', data: {videoData: this}};
+    }
+    if(!this.resizer) {
+      this.destroy();
+      return;
+    }
+    this.resizer.panHandler(event);
+  }
+
+  setPanMode(mode) {
+    this.resizer.setPanMode(mode);
+  }
+
+  setVideoFloat(videoFloat) {
+    this.resizer.setVideoFloat(videoFloat);
+  }
+
   restoreAr(){
     this.resizer.restore();
   }
@@ -103,8 +155,15 @@ class VideoData {
     this.resizer.setStretchMode(stretchMode);
   }
 
-  zoomStep(step){
-    this.resizer.zoomStep();
+  setZoom(zoomLevel, no_announce){
+    this.resizer.setZoom(zoomLevel, no_announce);
   }
 
+  zoomStep(step){
+    this.resizer.zoomStep(step);
+  }
+
+  announceZoom(scale){
+    this.pageInfo.announceZoom(scale);
+  }
 }