+++ js/modules/PageInfo.js
@@ -2,12 +2,11 @@ if(Debug.debug)
   console.log("Loading: PageInfo.js");
 
 class PageInfo {
-  constructor(comms){
-    this.keybinds = new Keybinds(this);
-    this.keybinds.setup();
+  constructor(comms, settings){
     this.hasVideos = false;
     this.siteDisabled = false;
     this.videos = [];
+    this.settings = settings;
 
     this.lastUrl = window.location.href;
 
@@ -16,15 +15,30 @@ class PageInfo {
 
     if(comms){ 
       this.comms = comms;
-      if(this.videos.length > 0){
-        comms.registerVideo();
-      }
+    }
+
+    if(this.videos.length > 0){
+      console.log("registering video")
+      comms.registerVideo();
+    }
+
+    this.currentZoomScale = 1;
+  }
+
+  destroy() {
+    if(Debug.debug || Debug.init){
+      console.log("[PageInfo::destroy] destroying all videos!")
+    }
+    if(this.rescanTimer){
+      clearTimeout(this.rescanTimer);
+    }
+    for (var video of this.videos) {
+      video.destroy();
     }
   }
 
-
-  reset(){
-    for(video of this.videos) {
+  reset() {
+    for(var video of this.videos) {
       video.destroy();
     }
     this.rescan(RescanReason.MANUAL);
@@ -82,7 +96,7 @@ class PageInfo {
         if(Debug.debug && Debug.periodic && Debug.videoRescan){
           console.log("[PageInfo::rescan] found new video candidate:", video)
         }
-        v = new VideoData(video);
+        v = new VideoData(video, this.settings, this);
         // console.log("[PageInfo::rescan] v is:", v)
         // debugger;
         v.initArDetection();
@@ -110,7 +124,6 @@ class PageInfo {
     this.videos = this.videos.filter( vid => vid.destroyed === false);
   }
 
-
   scheduleRescan(rescanReason){
     if(rescanReason != RescanReason.PERIODIC){
       this.rescan(rescanReason);
@@ -129,7 +142,7 @@ class PageInfo {
         ths.rescanTimer = null;
         ths.rescan(rr);
         ths = null;
-      }, rescanReason === ExtensionConf.pageInfo.timeouts.rescan, RescanReason.PERIODIC)
+      }, rescanReason === this.settings.active.pageInfo.timeouts.rescan, RescanReason.PERIODIC)
     } catch(e) {
       if(Debug.debug){
         console.log("[PageInfo::scheduleRescan] scheduling rescan failed. Here's why:",e)
@@ -149,7 +162,7 @@ class PageInfo {
       ths.rescanTimer = null;
       ths.ghettoUrlCheck();
       ths = null;
-    }, ExtensionConf.pageInfo.timeouts.urlCheck)
+    }, this.settings.active.pageInfo.timeouts.urlCheck)
     }catch(e){
       if(Debug.debug){
         console.log("[PageInfo::scheduleUrlCheck] scheduling URL check failed. Here's why:",e)
@@ -212,12 +225,31 @@ class PageInfo {
     if(ar !== 'auto') {
       this.stopArDetection();
     }
+
     // TODO: find a way to only change aspect ratio for one video
-    for(var vd of this.videos){
-      vd.setAr(ar)
+    if (ar === 'reset') {
+      for (var vd of this.videos) {
+        vd.resetAr();
+      }
+    } else {
+      for (var vd of this.videos) {
+        vd.setAr(ar)
+      }
     }
   }
   
+  setVideoFloat(videoFloat) {
+    for(var vd of this.videos) {
+      vd.setVideoFloat(videoFloat)
+    }
+  }
+
+  setPanMode(mode) {
+    for(var vd of this.videos) {
+      vd.setPanMode(mode);
+    }
+  }
+
   restoreAr() {
     for(var vd of this.videos){
       vd.restoreAr()
@@ -232,10 +264,29 @@ class PageInfo {
     }
   }
 
+  setZoom(zoomLevel, no_announce) {
+    for(var vd of this.videos) {
+      vd.setZoom(zoomLevel, no_announce);
+    }
+  }
+
   zoomStep(step){
     for(var vd of this.videos){
       vd.zoomStep(step);
     }
+  }
+
+  announceZoom(scale) {
+    if (this.announceZoomTimeout) {
+      clearTimeout(this.announceZoom);
+    }
+    this.currentZoomScale = scale;
+    const ths = this;
+    this.announceZoomTimeout = setTimeout(() => ths.comms.announceZoom(scale), this.settings.active.zoom.announceDebounce);
+  }
+
+  requestCurrentZoom() {
+    this.comms.announceZoom(this.currentZoomScale);
   }
 }