+++ js/modules/Resizer.js
@@ -13,7 +13,7 @@ class Resizer {
   constructor(videoData){
     this.conf = videoData;
     this.video = videoData.video;
-
+    this.settings = videoData.settings;
 
     this.scaler = new Scaler(this.conf);
     this.stretcher = new Stretcher(this.conf); 
@@ -22,18 +22,36 @@ class Resizer {
     // load up default values
     this.correctedVideoDimensions = {};
     this.currentCss = {};
+    this.currentStyleString = "";
+    this.currentCssValidFor = {};
 
     // restore watchdog. While true, applyCss() tries to re-apply new css until this value becomes false again
     // value becomes false when width and height of <video> tag match with what we want to set. Only necessary when
     // calling _res_restore() for some weird reason.
     this.restore_wd = false;
 
-    this.lastAr = {type: 'original'};
+    // CSS watcher will trigger _very_ often for this many iterations
+    this.cssWatcherIncreasedFrequencyCounter = 0;
+
+    
+    this.lastAr = this.settings.getDefaultAr();                 // this is the aspect ratio we start with
+    this.videoFloat = this.settings.getDefaultVideoAlignment(); // this is initial video alignment
     this.destroyed = false;
+
+    this.resizerId = (Math.random(99)*100).toFixed(0);
+
+    if (this.settings.active.pan) {
+      console.log("can pan:", this.settings.active.miscFullscreenSettings.mousePan.enabled, "(default:", this.settings.active.miscFullscreenSettings.mousePan.enabled, ")")
+      this.canPan = this.settings.active.miscFullscreenSettings.mousePan.enabled;
+    } else {
+      this.canPan = false;
+    }
   }
   
   start(){
-    this.startCssWatcher();
+    if(!this.destroyed) {
+      this.startCssWatcher();
+    }
   }
 
   stop(){
@@ -41,14 +59,23 @@ class Resizer {
   }
 
   destroy(){
+    if(Debug.debug || Debug.init){
+      console.log(`[Resizer::destroy] <rid:${this.resizerId}> received destroy command.`);
+    }
     this.destroyed = true;
     this.stopCssWatcher();
   }
 
 
   setAr(ar, lastAr){
+    if (this.destroyed) {
+      return;
+    }
+    this.startCssWatcher();
+    this.cssWatcherIncreasedFrequencyCounter = 20;
+    
     if(Debug.debug){
-      console.log('[Resizer::setAr] trying to set ar. New ar:', ar)
+      console.log('[Resizer::setAr] <rid:'+this.resizerId+'> trying to set ar. New ar:', ar)
     }
 
     if(lastAr) {
@@ -66,7 +93,7 @@ class Resizer {
       this.videoData.destroy();
     }
 
-    // pause AR on basic stretch, unpause when using other mdoes
+    // // pause AR on basic stretch, unpause when using other mdoes
     // fir sine reason unpause doesn't unpause. investigate that later
     // if (this.stretcher.mode === StretchMode.BASIC) {
     //   this.conf.arDetector.pause();
@@ -80,7 +107,7 @@ class Resizer {
 
       if(! stretchFactors || stretchFactors.error){
         if(Debug.debug){
-          console.log("[Resizer::setAr] failed to set AR due to problem with calculating crop. Error:", (stretchFactors ? stretchFactors.error : stretchFactors));
+          console.log("[Resizer::setAr] <rid:"+this.resizerId+"> failed to set AR due to problem with calculating crop. Error:", (stretchFactors ? stretchFactors.error : stretchFactors));
         }
         if(stretchFactors.error === 'no_video'){
           this.conf.destroy();
@@ -121,6 +148,24 @@ class Resizer {
     this.restore();
   }
 
+  panHandler(event) {
+    // console.log("this.conf.canPan:", this.conf.canPan)
+    if (this.canPan) {
+      // console.log("event?", event)
+      // console.log("this?", this)
+
+      if(!this.conf.player || !this.conf.player.element) {
+        return;
+      }
+      const player = this.conf.player.element;
+
+      const relativeX = (event.pageX - player.offsetLeft) / player.offsetWidth;
+      const relativeY = (event.pageY - player.offsetTop) / player.offsetHeight;
+      
+      this.setPan(relativeX, relativeY);
+    }
+  }
+
   setPan(relativeMousePosX, relativeMousePosY){
     // relativeMousePos[X|Y] - on scale from 0 to 1, how close is the mouse to player edges. 
     // use these values: top, left: 0, bottom, right: 1
@@ -128,18 +173,63 @@ class Resizer {
       this.pan = {};
     }
 
-    this.pan.relativeOffsetX = relativeMousePosX*2.5 - 1.25;
-    this.pan.relativeOffsetY = relativeMousePosY*2.5 - 1.25;
+    this.pan.relativeOffsetX = -(relativeMousePosX * 1.1) + 0.55;
+    this.pan.relativeOffsetY = -(relativeMousePosY * 1.1) + 0.55;
+
+    // if(Debug.debug){
+    //   console.log("[Resizer::setPan] relative cursor pos:", relativeMousePosX, ",",relativeMousePosY, " | new pan obj:", this.pan)
+    // }
+    this.restore();
+  }
+
+  setVideoFloat(videoFloat) {
+    this.videoFloat = videoFloat;
+    this.restore();
   }
 
   startCssWatcher(){
-    // this.haltCssWatcher = false;
-    if(!this.cssWatcherTimeout){
-      // if(Debug.debug)
-        // console.log("[Resizer.js] STARTING CSS WATCHER")
-  
-      // this.cssWatcherTimeout = setInterval(this.cssWatcher, 200, this);
+    if(Debug.debug) {
+      console.log("[Resizer.js::startCssWatcher] starting css watcher. Is resizer destroyed?", this.destroyed);
     }
+    if (this.destroyed) {
+      return;
+    }
+
+    // this.haltCssWatcher = false;
+    if(!this.cssWatcherTimer){
+      this.scheduleCssWatcher(1);
+    } else {
+      clearTimeout(this.cssWatcherTimer);
+      this.scheduleCssWatcher(1);
+    }
+  }
+  
+  scheduleCssWatcher(timeout, force_reset) {
+    if (this.destroyed) {
+      return;
+    }
+
+    if(timeout === undefined) {
+      console.log("?")
+      this.cssCheck(); // no timeout = one-off
+      return;
+    }
+
+    if(this.cssWatcherTimeout) {
+      clearTimeout(this.cssWatcherTimer);
+    }
+
+    var ths = this;
+    this.cssWatcherTimer = setTimeout(function () {
+        ths.cssWatcherTimer = null;
+        try {
+          ths.cssCheck();
+        } catch (e) {
+          if(Debug.debug) {
+            console.log("[Resizer.js::scheduleCssWatcher] Css check failed. Error:", e);
+          }
+        }
+      }, timeout);
   }
 
   stopCssWatcher() {
@@ -150,7 +240,7 @@ class Resizer {
 
   restore() {
     if(Debug.debug){
-      console.log("[Resizer::restore] attempting to restore aspect ratio. this & settings:", {'this': this, "settings": Settings} );
+      console.log("[Resizer::restore] <rid:"+this.resizerId+"> attempting to restore aspect ratio. this & settings:", {'this': this, "settings": this.settings} );
     }
     
     // this is true until we verify that css has actually been applied
@@ -171,10 +261,24 @@ class Resizer {
     this.setAr('reset');
   }
 
+  setPanMode(mode) {
+    if (mode === 'enable') {
+      this.canPan = true;
+    } else if (mode === 'disable') {
+      this.canPan = false;
+    } else if (mode === 'toggle') {
+      this.canPan = !this.canPan;
+    }
+  }
+
   resetPan(){
     this.pan = undefined;
   }
 
+  setZoom(zoomLevel, no_announce) {
+    this.zoom.setZoom(zoomLevel, no_announce);
+  }
+  
   zoomStep(step){
     this.zoom.zoomStep(step);
   }
@@ -198,23 +302,36 @@ class Resizer {
 
   computeOffsets(stretchFactors){
 
-    if(Debug.debug)
-      console.log("[Resizer::_res_computeOffsets] video will be aligned to ", ExtensionConf.miscFullscreenSettings.videoFloat);
-  
+    if (Debug.debug) {
+      console.log("[Resizer::_res_computeOffsets] <rid:"+this.resizerId+"> video will be aligned to ", this.settings.active.miscFullscreenSettings.videoFloat);
+    }
+
     var actualWidth = this.conf.video.offsetWidth * stretchFactors.xFactor;
     var actualHeight = this.conf.video.offsetHeight * stretchFactors.yFactor;
+
+    var wdiff = actualWidth - this.conf.player.dimensions.width;
+    var hdiff = actualHeight - this.conf.player.dimensions.height;
 
     var translate = {x: 0, y: 0};
 
     if (this.pan) {
-      // todo: calculate translate
+      // don't offset when video is smaller than player
+      if(wdiff < 0 && hdiff < 0) {
+        return translate;
+      }
+      translate.x = wdiff * this.pan.relativeOffsetX / this.zoom.scale;
+      translate.y = hdiff * this.pan.relativeOffsetY / this.zoom.scale;
     } else {
-      if (ExtensionConf.miscFullscreenSettings.videoFloat == "left") {
-        translate.x = (this.conf.player.dimensions.width - actualWidth) * -0.5;
+      if (this.videoFloat == "left") {
+        translate.x = wdiff * 0.5;
       }
-      else if (ExtensionConf.miscFullscreenSettings.videoFloat == "right") {
-        translate.x = (this.conf.player.dimensions.width - actualWidth) * 0.5;
+      else if (this.videoFloat == "right") {
+        translate.x = wdiff * -0.5;
       }
+    }
+
+    if(Debug.debug) {
+      console.log("[Resizer::_res_computeOffsets] <rid:"+this.resizerId+"> calculated offsets:", translate);
     }
 
     return translate; 
@@ -223,8 +340,10 @@ class Resizer {
   applyCss(stretchFactors, translate){
 
     if (! this.video) {
-      if(Debug.debug)
-        console.log("[Resizer::_res_applyCss] Video went missing, doing nothing.");
+      if(Debug.debug) {
+        console.log("[Resizer::_res_applyCss] <rid:"+this.resizerId+"> Video went missing, doing nothing.");
+      }
+
       this.conf.destroy();
       return;
     }
@@ -271,12 +390,15 @@ class Resizer {
 
   setStyleString (styleString, count = 0) {
     this.video.setAttribute("style", styleString);
+
+    this.currentStyleString = this.video.getAttribute('style');
+    this.currentCssValidFor = this.conf.player.dimensions;
     
     if(this.restore_wd){
   
       if(! this.video){
         if(Debug.debug)
-          console.log("[Resizer::_res_setStyleString] Video element went missing, nothing to do here.")
+          console.log("[Resizer::_res_setStyleString] <rid:"+this.resizerId+"> Video element went missing, nothing to do here.")
         return;
       }
       
@@ -288,8 +410,8 @@ class Resizer {
       //   if(Debug.debug)
       //     console.log("[Resizer::_res_setStyleString] Style string not set ???");
         
-      //   if(count < ExtensionConf.resizer.setStyleString.maxRetries){
-      //     setTimeout( this.setStyleString, ExtensionConf.resizer.setStyleString.retryTimeout, count + 1);
+      //   if(count < settings.active.resizer.setStyleString.maxRetries){
+      //     setTimeout( this.setStyleString, settings.active.resizer.setStyleString.retryTimeout, count + 1);
       //   }
       //   else if(Debug.debug){
       //     console.log("[Resizer::_res_setStyleString] we give up. css string won't be set");
@@ -301,78 +423,48 @@ class Resizer {
     }
     else{
       if(Debug.debug)
-        console.log("[Resizer::_res_setStyleString] css applied. Style string:", styleString);
+        console.log("[Resizer::_res_setStyleString] <rid:"+this.resizerId+"> css applied. Style string:", styleString);
     }
   }
 
-  cssWatcher(ths){
+  cssCheck(){
     // this means we haven't set our CSS yet, or that we changed video.
-    if(! ths.currentCss.top)
-      return;
+    // if(! this.currentCss.tranform) {
+    //   this.scheduleCssWatcher(200);      
+    //   return;
+    // }
     
     // this means video went missing. videoData will be re-initialized when the next video is found
-    if(! ths.video){
-      ths.conf.destroy();
+    if(! this.video){
+      if(Debug.debug) {
+        console.log("[Resizer::cssCheck] <rid:"+this.resizerId+"> no video detecting, issuing destroy command");
+      }
+      this.conf.destroy();
       return;
     }
-    console.log("css watcher running. video?", ths.video)
     
-    // // our current css is fucky? Null, undefined and 0 are invalid values.
-    // if(! GlobalVars.currentCss.width || ! GlobalVars.currentCss.height )
-    //   return;
+    if(this.destroyed) {
+      if(Debug.debug) {
+        console.log("[Resizer::cssCheck] <rid:"+this.resizerId+"> destroyed flag is set, we shouldnt be running");
+      }
+      this.stopCssWatcher();
+      return;
+    }
+
+    var styleString = this.video.getAttribute('style');
 
     // first, a quick test:
-    if (ths.currentVideoSettings.validFor == ths.conf.player.dimensions ){
-      if (ths.currentVideoSettings.videoWidth != ths.video.offsetWidth  ||
-          ths.currentVideoSettings.videoHeight != ths.video.offsetHeight){
-        ths.restore();
-        return;
-      }
+    // if (this.currentVideoSettings.validFor == this.conf.player.dimensions ){
+    if (this.currentStyleString !== styleString){
+      this.restore();
+      this.scheduleCssWatcher(10);
+      return;
     }
-  
-    var styleArrayStr = ths.video.getAttribute('style');
-    
-    if (styleArrayStr){
-      var styleArray = styleArrayStr.split(";");
-  
-      var stuffChecked = 0;
-      var stuffToCheck = 2;
-      
-      for(var i in styleArray){
-        styleArray[i] = styleArray[i].trim();
-        
-        if (styleArray[i].startsWith("top:")){
-          // don't force css restore if currentCss.top is not defined
-          if(ths.currentCss.top && styleArray[i] != ths.currentCss.top){
-            if(Debug.debug){
-              console.log("[Resizer::_res_antiCssOverride] SOMEBODY TOUCHED MA SPAGHETT (our CSS got overriden, restoring our css)");
-              console.log("[Resizer::_res_antiCssOverride] MA SPAGHETT: top:", ths.currentCss.top, "left:", ths.currentCss.left, "thing that touched ma spaghett", styleArrayStr);
-            }
-            ths.restore();
-            return;
-          }
-          stuffChecked++;
-        }
-        else if(styleArray[i].startsWith("left:")){
-          // don't force css restore if currentCss.left is not defined        
-          if(ths.currentCss.left && styleArray[i] != ths.currentCss.left){
-            if(Debug.debug){
-              console.log("[Resizer::_res_antiCssOverride] SOMEBODY TOUCHED MA SPAGHETT (our CSS got overriden, restoring our css)");
-              console.log("[Resizer::_res_antiCssOverride] MA SPAGHETT: width:", ths.currentCss.width, "height:", ths.currentCss.height, "thing that touched ma spaghett", styleArrayStr);            
-            }
-            ths.restore();
-            return;
-          }
-          stuffChecked++;
-        }
-        
-        if(stuffChecked == stuffToCheck){
-    //         if(Debug.debug){
-    //           console.log("[Resizer::_res_antiCssOverride] My spaghett rests untouched. (nobody overrode our CSS, doing nothing)");
-    //         }
-          return;
-        }
-      }
+    if (this.cssWatcherIncreasedFrequencyCounter > 0) {
+      --this.cssWatcherIncreasedFrequencyCounter;
+      this.scheduleCssWatcher(20);
+    } else {
+      this.scheduleCssWatcher(1000);    
     }
   }
 }